This commit was manufactured by cvs2svn to create tag 'V1_0_5_FREEZE'.
git-svn-id: svn://anonsvn.mit.edu/krb5/tags/V1_0_5_FREEZE@10386 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/kadmin/passwd/kpasswd.c
@@ -137,7 +137,7 @@ kpasswd(context, argc, argv)
 
     /* if either krb5_cc failed check the passwd file */
     if (code != 0) {
-      pw = getpwuid((int) getuid());
+      pw = getpwuid( getuid());
       if (pw == NULL) {
 	com_err(whoami, 0, string_text(KPW_STR_NOT_IN_PASSWD_FILE));
 	return(MISC_EXIT_STATUS);
